[CLEANUP] Move components to glimmer
The original code is moved from `ember-htmlbars` to `ember-glimmer` and
symlinked from `ember-htmlbars`

Fixes #14020
